IMAGE backend/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
pillow
sqlite3
python-jose
openai